feat: automatic Compodoc generation by one command
Added an extra schematic which build either all available projects or a
specific set of provided projects. It will execute the compodoc build
command and build documentation for those projects. It will also create
an extra website which contains a clear overview of all projects (lib
and apps), where you can navigate to the documentation of other libs.

Signed-off-by: matthiasgekiere <matthias.gekiere@rydoo.com>